Added Project 01 to the repository
Recoverability In AWS

In this project a highly available solutions to common use cases is created. A Multi-AvailabilityZone, Multi-Region database and show how to use it in multiple geographically separate AWS regions is built. A website hosting solution that is versioned so that any data destruction and accidents can be quickly and easily undone is also built.

- Various screenshots as denoted in the instructions
- Database command line logs from the various interactions with the databases of this project
- A written description of the RTO and RPO related to this project as described within the instructions